stats display on homepage,  support for immutable status of pymtAcc, issues fix
- fix issue of demo expense  api, to getlist by expense date not updatedOn
- added stats api demo
- page section for stats, service and route handler

+++ src/pages/pymt-accounts/components/view-account/account-list.tsx
@@ -1,7 +1,7 @@
 import { FunctionComponent, useState } from "react";
 import { useActionData, useLoaderData, useSubmit } from "react-router-dom";
-import AccountItemCard from "./account-item-view";
-import { PymtAccountFields, RouteHandlerResponse } from "../../services";
+import { AccountItemCard } from "./account-item-view";
+import { PymtAccountFields, PymtAccStatus, RouteHandlerResponse } from "../../services";
 import { Animated, ConfirmDialog } from "../../../../components";
 import { getFullPath } from "../../../root";
 import ReactMarkdown from "react-markdown";
@@ -18,9 +18,11 @@ const AccountList: FunctionComponent = () => {
     const onDeleteConfirmHandler = () => {
         if (loaderData.type === "success" && !auth.readOnly) {
             const deletingPymtAcc = loaderData.data.find(acc => acc.id === deletingAccountId);
-            const data: any = { ...deletingPymtAcc };
-            submit(data, { action: getFullPath("pymtAccountsRoot"), method: "delete", encType: "application/json" });
-            setDeletingAccountId("");
+            if (deletingPymtAcc && deletingPymtAcc.status === PymtAccStatus.Enable) {
+                const data: any = { ...deletingPymtAcc };
+                submit(data, { action: getFullPath("pymtAccountsRoot"), method: "delete", encType: "application/json" });
+                setDeletingAccountId("");
+            }
         }
     };
 
@@ -47,13 +49,14 @@ const AccountList: FunctionComponent = () => {
                 !errorMessage && !pymtAccList.length &&
                 <p className="title">There are no accounts</p>
             }
-            { pymtAccList.map(acc =>
-                <AccountItemCard
-                    key={ acc.id + "viewcard" }
-                    id={ acc.id + "viewcard" }
-                    details={ acc }
-                    onDeleteRequest={ setDeletingAccountId }
-                />)
+            {
+                pymtAccList.map(acc =>
+                    <AccountItemCard
+                        key={ acc.id + "viewcard" }
+                        id={ acc.id + "viewcard" }
+                        details={ acc }
+                        onDeleteRequest={ setDeletingAccountId }
+                    />)
             }
             <ConfirmDialog
                 id="delete-pymt-acc-confirm-dialog"
